acpi_tables: Swap out vm-memory for zerocopy
vm-memory is a little heavyweight for just the `as_slice()`
fn. Instead, `zerocopy` is a widely-used library with the ability to
derive an implementation of an `as_bytes()` fn which returns a
read-only byte-slice of a struct.

Also explicitly define the endianness of the multi-byte integers in
the Rsdp struct as little-endian (what ACPI expects), instead of
relying on the host order to be correct.

Signed-off-by: Tim Wawrzynczak <tim@rivosinc.com>