acpi_tables: Change output type of the Aml trait
Instead of directly writing to a Vec<u8>, it may be useful in other
environments to have a trait which is capable of accepting bytes.
There may be use-cases where a buffer has already been provided for
accepting ACPI tables, for example.

In addition, move all of the open-coding of AML opcodes into enums.

Signed-off-by: Tim Wawrzynczak <tim@rivosinc.com>