acpi_tables: Update to latest RQSC specs
The table definition for RQSC tables changed[0]. This commits updates
the RQSC implementation to match the new specs.

The new specs can be confusing to use, since some fields depends on
others. Sometime they should be set, sometimes not. To prevent mishaps,
enums are used to force the developer to be coherent in what he defines.

[0] https://github.com/riscv-non-isa/riscv-rqsc

Signed-off-by: Simon Lucido <simonl@rivosinc.com>